Changed game view into window. Added some basic add/remove functionality. Renamed datasource

+++ src/main/java/se/ifkgoteborg/stat/ui/SeasonView.java
@@ -4,11 +4,15 @@ import java.util.List;
 
 import se.ifkgoteborg.stat.controller.RegistrationDAO;
 import se.ifkgoteborg.stat.model.PlayedForClub;
-import se.ifkgoteborg.stat.model.Season;
+import se.ifkgoteborg.stat.model.SquadSeason;
+import se.ifkgoteborg.stat.ui.editor.PlayerEditor.EditorSavedEvent;
+import se.ifkgoteborg.stat.ui.editor.PlayerEditor.EditorSavedListener;
+import se.ifkgoteborg.stat.ui.editor.PlayerSeasonEditor;
 
 import com.vaadin.data.Property;
 import com.vaadin.data.Property.ValueChangeEvent;
 import com.vaadin.data.Property.ValueChangeListener;
+import com.vaadin.data.util.BeanItem;
 import com.vaadin.data.util.BeanItemContainer;
 import com.vaadin.ui.Button;
 import com.vaadin.ui.Button.ClickEvent;
@@ -23,54 +27,34 @@ public class SeasonView extends VerticalLayout {
 	ComboBox seasonCombo = new ComboBox("Säsonger");
 	Table table = new Table("Spelare vald säsong");
 	
-	Button addBtn = new Button("Lägg till");
-	Button removeBtn = new Button("Ta bort");
+	Button addBtn = new Button("Lägg till spelare");
+	Button editBtn = new Button("Redigera vald spelare");
+	Button removeBtn = new Button("Ta bort spelare");
 	Button saveBtn = new Button("Spara");
 	
-	BeanItemContainer<Season> bic = new BeanItemContainer<Season>(Season.class);
+	BeanItemContainer<SquadSeason> bic = new BeanItemContainer<SquadSeason>(SquadSeason.class);
 	BeanItemContainer<PlayedForClub> tableBic = new BeanItemContainer<PlayedForClub>(PlayedForClub.class);
 	
 	private final RegistrationDAO dao;
 	
 	private String[] cols = new String[]{"player.name", "squadNr", "importIndex"};
 	protected PlayedForClub selectedItem;
+	protected SquadSeason selectedSeason;
 	
 	public SeasonView(RegistrationDAO dao) {
 		this.dao = dao;
 		setupButtons();
 		loadSeasons();
-		seasonCombo.setContainerDataSource(bic);
-		seasonCombo.addListener(new ValueChangeListener() {
-			
-			@Override
-			public void valueChange(ValueChangeEvent event) {
-				Object value = event.getProperty().getValue();
-				if(value instanceof Season) {
-					loadSeasonIntoTable(((Season) value).getSquad());
-				}
-			}
-
-			
-		});
+		
+		
 		table.setContainerDataSource(tableBic);
-//		table.setTableFieldFactory(new TableFieldFactory() {
-//
-//            @Override
-//            public Field createField(Container container, Object itemId, Object propertyId, Component uiContext) {
-//                System.out.println(container.toString() + itemId.toString() + propertyId.toString() + uiContext.toString());
-//                String pid = (String) propertyId;
-//                if("squadNr".equals(pid)) {
-//                	return DefaultFieldFactory.get().createField(container, itemId, propertyId, uiContext);
-//                }
-//                return null;
-//            }
-//        });
+
 		tableBic.addNestedContainerProperty("player.name");
 		table.setColumnHeader("player.name", "Namn");   
 		table.setColumnHeader("squadNr", "Spelarnummer");
 		table.setColumnHeader("importIndex", "Indexplats");
-		//table.setColumnHeader("btn", "");
-		table.setEditable(true);
+
+		
 		table.setVisibleColumns(cols);
 		table.setSelectable(true);
 		table.setImmediate(true);
@@ -79,20 +63,38 @@ public class SeasonView extends VerticalLayout {
 		    public void valueChange(ValueChangeEvent event) {
 		    	selectedItem = (PlayedForClub) table.getValue();
 				removeBtn.setEnabled(true);
+				editBtn.setEnabled(true);
 		    }
 		});
 		
 		HorizontalLayout toolBar = new HorizontalLayout();
 		toolBar.addComponent(addBtn);
+		toolBar.addComponent(editBtn);
 		toolBar.addComponent(removeBtn);
 		
 		addComponent(seasonCombo);
 		addComponent(toolBar);
-		addComponent(table);
+		
 		
 		HorizontalLayout footerBar = new HorizontalLayout();
 		footerBar.addComponent(saveBtn);
 		addComponent(footerBar);
+		
+		seasonCombo.addListener(new ValueChangeListener() {
+			
+			@Override
+			public void valueChange(ValueChangeEvent event) {
+				Object value = event.getProperty().getValue();
+				if(value instanceof SquadSeason) {
+					selectedSeason = (SquadSeason) value;
+					System.out.println("Selected Season: " + value.toString());
+					addComponent(table);
+					loadSeasonIntoTable(((SquadSeason) value).getSquad());
+				}
+			}
+		});
+		
+		seasonCombo.setContainerDataSource(bic);
 	}
 
 	private void setupButtons() {
@@ -100,14 +102,35 @@ public class SeasonView extends VerticalLayout {
 			
 			@Override
 			public void buttonClick(ClickEvent event) {
-				tableBic.addBean(new PlayedForClub());
+				// tableBic.addBean(new PlayedForClub());
+				// open "new" window.
+                
+                PlayerSeasonEditor psEditor = new PlayerSeasonEditor(new PlayedForClub(), selectedSeason, SeasonView.this.dao, SeasonView.this);
+                getApplication().getMainWindow().addWindow(psEditor);
 			}
 		});
+		
+		editBtn.addListener(new ClickListener() {
+			
+			@Override
+			public void buttonClick(ClickEvent event) {				
+				// open "edit" window.
+				if(selectedSeason == null || selectedItem == null) {
+					return;
+				}
+				PlayerSeasonEditor psEditor = new PlayerSeasonEditor((PlayedForClub) selectedItem, selectedSeason, SeasonView.this.dao, SeasonView.this);
+				getApplication().getMainWindow().addWindow(psEditor);
+			}
+		});
+		editBtn.setEnabled(false);
 		
 		removeBtn.addListener(new ClickListener() {
 			
 			@Override
 			public void buttonClick(ClickEvent event) {
+				if(selectedSeason == null || selectedItem == null) {
+					return;
+				}
 				tableBic.removeItem(selectedItem);
 				dao.removePlayedForClub(selectedItem);
 				selectedItem = null;	
@@ -115,6 +138,7 @@ public class SeasonView extends VerticalLayout {
 				table.commit();
 			}
 		});
+		removeBtn.setEnabled(false);
 	}
 
 	private void loadSeasonIntoTable(List<PlayedForClub> squad) {
@@ -122,35 +146,34 @@ public class SeasonView extends VerticalLayout {
 		int i = 0;
 		
 		for(PlayedForClub pfc : squad) {
-			
-//			Integer itemId = new Integer(i);
-//			BeanItem<PlayedForClub> bi = new BeanItem<PlayedForClub>(pfc);
-//			Button buttonField = new Button("Ändra");
-//		    buttonField.setData(itemId);
-//		    buttonField.addListener(new Button.ClickListener() {
-//		        public void buttonClick(ClickEvent event) {
-//		            // Get the item identifier from the user-defined data.
-//		            Integer itemId = (Integer)event.getButton().getData();
-//		            getWindow().showNotification("Link "+
-//		                                   itemId.intValue()+" clicked.");
-//		        } 
-//		    });
-//		    buttonField.addStyleName("link");
-//			bi.addItemProperty("btn", buttonField);
-//						
+		
 			tableBic.addBean(pfc);	
 			i++;
-		}	
-		//tableBic.addNestedContainerProperty("btn");
+		}
 	}
 	
 	private void loadSeasons() {
-		List<Season> seasons = dao.getSeasons();
+		List<SquadSeason> seasons = dao.getSeasons();
 		
 		bic.removeAllItems();
-		for(Season s : seasons) {
+		for(SquadSeason s : seasons) {
 			bic.addBean(s);
 		}
+	}
+
+	public void refresh() {
+		tableBic.removeAllItems();
+		loadSeasons();
+		System.out.println("Refreshed SeasonView");
+	}
+
+	public void reloadTable() {	
+		
+		// E.g. reload from database.
+		selectedSeason = dao.getSquadSeason(selectedSeason.getId());
+		
+		tableBic.removeAllItems();
+		loadSeasonIntoTable(selectedSeason.getSquad());
 	}
 	
 }
